# api-doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "bloggorras Api Documentation",
    "description": "bloggorras Api Documentation",
    "license": {
      "name": "Specify your license",
      "url": "https://yourlicense.com"
    },
    "x-logo": {
      "url": "logo.png",
      "description": "bloggorras Logo",
      "altText": "bloggorras Logo",
      "backgroundColor": "#483D8B"
    },
    "contact": {
      "name": "bloggorras API Support",
      "url": "https://bloggorras.com/info/support",
      "email": "support@bloggorras.com"
    },
    "termsOfService": "https://bloggorras.com/info/terms_and_conditions"
  },
  "servers": [
    {
      "url": "http://localhost:8060/api",
      "description": "Api base url"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "User Login",
        "description": "Authenticate user and get user login data",
        "operationId": "UserLogin",
        
        "tags": [
          "Auth"
        ],
        
        "requestBody": {
          "description": "User login credentials",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginData"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthorized"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/register": {
      "post": {
        "summary": "User Registration",
        "description": "Agregar nuevo",
        "operationId": "UserRegister",
        
        "tags": [
          "Auth"
        ],
        
        "requestBody": {
          "description": "User Registration",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginData"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/account/edit": {
      "put": {
        "summary": "Edit User Account Details",
        "description": "Mi cuenta",
        "operationId": "AccountEdit",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Account"
        ],
        
        "requestBody": {
          "description": "Edit User Account Details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountEdit"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/account": {
      "get": {
        "summary": "User Account Details",
        "description": "Mi cuenta",
        "operationId": "Account",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Account"
        ],
        
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UserRegister": {
        "type": "object",
        "properties": {
          "name": {
            "description": "Name",
            "type": "string"
          },
          "email": {
            "description": "Email",
            "type": "string"
          },
          "password": {
            "description": "Password",
            "type": "string"
          },
          "confirm_password": {
            "description": "Confirm Password",
            "type": "string"
          },
          "phone": {
            "description": "Phone",
            "type": "string"
          },
          "photo": {
            "description": "Photo",
            "type": "string"
          }
        }
      },
      "AccountEdit": {
        "type": "object",
        "properties": {
          "name": {
            "description": "Name",
            "type": "string"
          },
          "phone": {
            "description": "Phone",
            "type": "string"
          },
          "photo": {
            "description": "Photo",
            "type": "string"
          },
          "user_role_id": {
            "description": "User Role Id",
            "type": "string"
          }
        }
      },
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "name": {
            "description": "Name",
            "type": "string"
          },
          "email": {
            "description": "Email",
            "type": "string"
          },
          "phone": {
            "description": "Phone",
            "type": "string"
          },
          "user_role_id": {
            "description": "User Role Id",
            "type": "string"
          }
        }
      },
...
      "UserLogin": {
        "type": "object",
        "properties": {
          "username": {
            "description": "Username or Email",
            "type": "string"
          },
          "password": {
            "description": "User Password",
            "type": "string"
          }
        }
      },
      "LoginData": {
        "type": "object",
        "properties": {
          "token": {
            "description": "User json web token(jwt)",
            "type": "string"
          },
          "user": {
            "description": "User Data",
            "type": "object"
          }
        }
      },
      "UnAuthenticated": {
        "type": "object",
        "properties": {
          "code": {
            "description": "401",
            "type": "integer"
          },
          "message": {
            "description": "UnAuthenticated",
            "type": "string"
          }
        }
      },
      "UnAuthorized": {
        "type": "object",
        "properties": {
          "code": {
            "description": "403",
            "type": "integer"
          },
          "message": {
            "description": "UnAuthorized",
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Http Status Code",
            "type": "integer"
          },
          "message": {
            "description": "Unable to complete request",
            "type": "string"
          }
        }
      }
    }
  }
}